+++ R/backend_api-SequentialFutureBackend-class.R
@@ -72,6 +72,8 @@ launchFuture.SequentialFutureBackend <- function(backend, future, ...) {
   split <- backend[["split"]]
   if (!is.null(split)) data[["capture"]][["split"]] <- split
 
+  future[["backend"]] <- backend
+
   ## Inherit 'earlySignal' from backend?
   earlySignal <- backend[["earlySignal"]]
   if (!is.null(earlySignal)) future[["earlySignal"]] <- earlySignal