+++ R/protected_api-signalConditions.R
@@ -102,7 +102,9 @@ signalConditions <- function(future, include = "condition", exclude = NULL, resi
       source <- attr(session_uuid, "source")
       host <- source[["host"]]
       pid <- source[["pid"]]
-      msg <- sprintf("A future ('%s') of class %s was interrupted at %s, while running on %s (pid %s)", label, class(future)[1], format(when, format = "%FT%T"), sQuote(host), pid)
+      msg <- sprintf("A future ('%s') of class %s was interrupted at %s, while running on %s (pid %s). The future was reset to a vanilla, lazy future, which can be re-evaluted again", label, class(future)[1], format(when, format = "%FT%T"), sQuote(host), pid)
+      backend <- future[["backend"]]
+      if (!is.null(backend)) future <- resetFuture(backend, future)
       warning(FutureInterruptWarning(msg, future = future))
     } else if (inherits(condition, "warning")) {
       warning(condition)